# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Prospector AtlasGR API
  version: "1.0.0"
  description: |
    Documentação da API do Prospector AtlasGR (CRM comercial + prospecção + IA).

    **Autenticação**: a maioria das rotas exige uma sessão válida do Better Auth,
    enviada via cookie de sessão (`better-auth.session_token`) definido por
    `POST /api/auth/sign-in/email` ou pelo fluxo social (`/api/auth/sign-in/social`).
    As rotas `/api/auth/*` são geridas pela biblioteca `better-auth` e não são
    implementadas manualmente neste projeto — os schemas de request abaixo
    (`SignUpEmailRequest`, `SignInEmailRequest`) refletem os campos aceitos pela
    lib, mas o contrato completo (inclusive `/api/auth/sign-in/social`,
    `/api/auth/sign-out`, `/api/auth/get-session`, etc.) é o padrão do Better Auth.

    **Multi-tenant**: toda rota autenticada (exceto `/api/auth/*`) roda sob
    `requireTenant`, que resolve o `organizationId` da sessão. Recursos operacionais
    são filtrados pelo tenant. O catálogo público oficial de Market Intelligence é
    global e somente leitura para tenants autenticados; ele nunca é misturado à
    entidade `Company` do CRM.

    **Envelope de resposta**: a maior parte dos endpoints devolve
    `{ "success": true, "data": ... }` (e opcionalmente `meta` para paginação) em
    caso de sucesso, ou `{ "success": false, "error": "..." }` em caso de falha.
    Alguns endpoints legados de `/api/intelligence` fogem do padrão e devolvem
...

...
  contact:
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Lista notificações do usuário autenticado
      parameters:
        - name: unread
          in: query
          schema: { type: string, enum: ['1', 'true'] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { type: object } }
                      unread: { type: integer }
  /notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Notifications]
      summary: Marca uma notificação como lida
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /notifications/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    delete:
      tags: [Notifications]
      summary: Remove uma notificação
      responses:
        '204':
          description: Removida
        '404':
          $ref: '#/components/responses/NotFound'

  # ───────────────────────────── Feature Flags ─────────────────────────────
...
  /integrations/birth-voice/call/{leadId}:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Birth Voice]
      summary: Dispara uma ligação do SDR de voz para o lead informado
      responses:
        '202':
          description: Ligação iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '422':
          description: Lead sem telefone válido
        '409':
          description: Número em lista de bloqueio (opt-out)
        '503':
          description: Integração Birth Voices Hub não configurada
  /integrations/birth-voice/suppressions:
    get:
      tags: [Birth Voice]
      summary: Lista os números em bloqueio (opt-out) de discagem para o tenant
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
    post:
      tags: [Birth Voice]
      summary: Bloqueia manualmente um número (opt-out registrado por outro canal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
                reason: { type: string, maxLength: 500 }
                leadId: { type: string }
      responses:
        '201':
          description: Bloqueado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '422':
          description: Telefone fora de um formato brasileiro discável

  # ═══════════════════════════════════════════════════════════════════════════════════
  # Endpoints adicionados em 2026-08-15 (Onda 8, Agente 18) para fechar a deriva medida
  # entre este documento e as rotas reais (`server.ts` + routers). Ver tabela de deriva
  # no relatório da Onda 8 — 6 routers de negócio inteiros e ~13 endpoints avulsos
  # existiam em código sem nenhuma documentação aqui.
  # ═══════════════════════════════════════════════════════════════════════════════════

  # ───────────────────────────── CRM 360 (Cockpit CRM) ─────────────────────────────
...
  /bitrix/connections/{connectionId}/test:
    parameters:
      - name: connectionId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Bitrix24]
      summary: Testa a conexão Bitrix24 (somente ADMIN/GESTOR)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DataEnvelope' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /integrations/3cx/connections/{connectionId}/test:
    parameters:
      - name: connectionId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [3CX]
      summary: Testa a conexão 3CX (somente ADMIN/GESTOR)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DataEnvelope' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: |
        Cookie de sessão emitido pelo Better Auth no login. Rotas protegidas passam por
        `authenticateToken` (valida a sessão) e `requireTenant` (resolve o organizationId).
        Endpoints marcados com roles (ex. `[ADMIN, GESTOR]`) também exigem `requireRole`.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: Recurso não encontrado (ou não pertence ao tenant autenticado)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Papel do usuário não tem permissão suficiente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      description: |
        Envelope de erro padrão, emitido pelo middleware global
        `src/shared/middlewares/errorHandler.ts` (último middleware da cadeia — ver `server.ts`).
        `success`/`error` existem desde a primeira versão da API; `code`, `requestId` e `retryable`
        foram adicionados depois como extensão aditiva (nenhum campo antigo foi removido) para
        cobrir o checklist de contratos do produto. `details` só aparece em alguns casos
        (ex.: `issues` do Zod em erro de validação).
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error: { type: string, description: 'Mensagem de erro legível, já existente antes da extensão de contrato.' }
        message: { type: string, description: 'Reservado no envelope de resposta; não é populado hoje por `errorHandler.ts`.' }
        details:
          description: 'Detalhe estruturado do erro quando disponível (ex.: `issues` do ZodError).'
        code:
          type: string
          description: |
            Código estável de erro, derivado do status HTTP e do tipo de erro já lançado no
            código (ZodError, `AppError`, `Prisma.PrismaClientKnownRequestError`). Pensado para o
            consumidor decidir programaticamente sem parsear `error`.
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - GONE
            - UNSUPPORTED_MEDIA_TYPE
            - UNPROCESSABLE_ENTITY
            - RATE_LIMITED
            - INTERNAL_ERROR
            - BAD_GATEWAY
            - SERVICE_UNAVAILABLE
            - TIMEOUT
            - BAD_REQUEST
        requestId:
          type: string
          description: |
            Mesmo id de correlação devolvido no header `x-request-id` (ver
            `src/shared/middlewares/observability.ts`) — não é gerado de novo aqui, é o mesmo
            valor reaproveitado. Ausente quando o erro ocorre antes de `observabilityMiddleware`
            rodar na cadeia de middlewares (ex.: falha em rota de auth/webhook montada antes dele).
        retryable:
          type: boolean
          description: |
            `true` para 5xx e timeouts (408/504) — falha transitória do servidor/upstream, repetir
            a mesma requisição pode funcionar. `false` para os demais 4xx (validação, autenticação,
            autorização, conflito, não encontrado) — o pedido em si precisa mudar antes de repetir.
    DataEnvelope:
      type: object
      description: Envelope de resposta genérico usado pela maioria dos endpoints
      properties:
        success: { type: boolean, example: true }
        data: {}
        meta: { type: object, description: 'Presente em listagens paginadas (page, limit, total, etc.)' }
...